# p1-back-sync/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "API Sync To-do",
      "description": "API para gerenciar tarefas de To-Dos",
      "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
      "/allTasks": {
        "get": {
          "summary": "Recupera todas as tarefas",
          "responses": {
            "200": {
              "description": "Lista de tarefas recuperada com sucesso",
              "schema": {
                "type": "object",
                "properties": {
                  "tasks": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Task"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/createTask": {
        "post": {
          "summary": "Cria uma nova tarefa",
          "parameters": [
            {
              "name": "task",
              "in": "body",
              "description": "Dados da tarefa a ser criada",
              "required": true,
              "schema": {
                "$ref": "#/definitions/TaskInput"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Tarefa criada com sucesso"
            }
          }
        }
      },
      "/removeTask": {
        "delete": {
          "summary": "Remove uma tarefa",
          "parameters": [
            {
              "name": "id",
              "in": "query",
              "description": "ID da tarefa a ser removida",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Tarefa removida com sucesso"
            }
          }
        }
      },
      "/updateTask": {
        "put": {
          "summary": "Atualiza uma tarefa",
          "parameters": [
            {
              "name": "task",
              "in": "body",
              "description": "Dados da tarefa a ser atualizada",
              "required": true,
              "schema": {
                "$ref": "#/definitions/TaskUpdate"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Tarefa atualizada com sucesso"
            }
          }
        }
      }
    },
    "definitions": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "TaskInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "TaskUpdate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
  